chat 렌더링최적화 2nd commit
