+++ src/components/chat/MessageInput.js
@@ -1,17 +1,37 @@
 import React, { useRef } from "react";
 import MyIcon from "../../element/MyIcon";
 
-const MessageInput = ({ setSendMessage, sendMessageFN, sendMessage }) => {
+const MessageInput = ({ stompClient }) => {
   const refInput = useRef();
+  const authNoBearer = sessionStorage.getItem("token")?.split(" ")[1];
+  const roomId = sessionStorage.getItem("roomId");
+  let sendMessage=""
+
   const lineBreackInput = () => {
     console.log(refInput.current.value);
     // refInput.current.value=refInput.current.value+"&#10;"
   };
   const sendmMessegeByEnter = () => {
-    console.log("실행됨!");
+ 
+    sendMessage=refInput.current.value
     sendMessageFN();
     refInput.current.value = "";
   };
+  function sendMessageFN() {
+    try {
+      stompClient.send(
+        "/pub/chat/message",
+        {
+          Authorization: authNoBearer,
+        },
+        JSON.stringify({
+          message: sendMessage,
+          type: "TALK",
+          roomId: roomId,
+        })
+      );
+    } catch (error) {}
+  }
 
   React.useEffect(() => {
     const keyDownHandler = (event) => {
@@ -50,15 +70,14 @@ const MessageInput = ({ setSendMessage, sendMessageFN, sendMessage }) => {
         <div
           className="chat-input-incon-circle"
           onClick={() => {
+           sendMessage=refInput.current.value
             sendMessageFN();
-            setSendMessage(refInput.current.value)
             refInput.current.value = "";
           }}
         >
           <MyIcon sizePx={23} iconName={"arrow_upward"} />
         </div>
       </div>
-      {/* <button onClick={sendMessageFN}>메시지보내기</button> */}
     </div>
   );
 };
